Improve test structure by defining namespace package
- move parsers' test methods inside their own file
